third_party/blink/web_tests/fast/borders/border-radius-child-expected.txt

Test for bug 113591: https://bugs.webkit.org/show_bug.cgi?id=113591 REGRESSION (r145870): Can't get insertion point to appear in some input and textareas on wordpress

PASS successfullyParsed is true

TEST COMPLETE

PASS document.elementFromPoint(x + 5, y - 5).id is 'outerBox'
PASS document.elementFromPoint(x + 5, y + 5).id is 'innerBox'
PASS document.elementFromPoint(x + 5, y + 95).id is 'innerBox'
PASS document.elementFromPoint(x + 5, y + 105).id is 'outerBox'
PASS document.elementFromPoint(x + 95, y - 5).id is 'outerBox'
PASS document.elementFromPoint(x + 95, y + 5).id is 'innerBox'
PASS document.elementFromPoint(x + 95, y + 95).id is 'innerBox'
PASS document.elementFromPoint(x + 95, y + 105).id is 'outerBox'
PASS document.elementFromPoint(x + 105, y - 5).id is 'container'
PASS document.elementFromPoint(x + 105, y + 5).id is 'innerBox'
PASS document.elementFromPoint(x + 105, y + 95).id is 'innerBox'
PASS document.elementFromPoint(x + 105, y + 105).id is 'container'
PASS document.elementFromPoint(x + 195, y - 5).id is 'container'
PASS document.elementFromPoint(x + 195, y + 5).id is 'innerBox'
PASS document.elementFromPoint(x + 195, y + 95).id is 'innerBox'
PASS document.elementFromPoint(x + 195, y + 105).id is 'container'
PASS successfullyParsed is true

TEST COMPLETE
